@@ -4,21 +4,130 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "loki"
-version = "0.0.1"
+version = "0.2.0"
 description = "Implementation of 'LoKI: Low-damage Knowledge Implanting of Large Language Models'"
 readme = "README.md"
-requires-python = ">=3.10"
-license = { file = "LICENSE" } # 或使用文件路径 license = {file = "LICENSE"}
+requires-python = ">=3.12"
+license = { file = "LICENSE" }
 dependencies = [
-    "transformers",
-    "huggingface_hub[hf_xet]",
-    "accelerate",
-    "peft",
-    "h5py",
-    "safetensors",
-    "numpy",
-    "datasets",
-    "tqdm",
-    "torch",
-    "captum",
+    "transformers>=4.36.0",
+    "huggingface_hub[hf_xet]>=0.19.0",
+    "accelerate>=0.25.0",
+    "peft>=0.7.0",
+    "h5py>=3.10.0",
+    "safetensors>=0.4.0",
+    "numpy>=1.24.0",
+    "datasets>=2.16.0",
+    "tqdm>=4.66.0",
+    "torch>=2.1.0",
+    "captum>=0.7.0",
 ]
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=7.4.0",
+    "pytest-cov>=4.1.0",
+    "black>=23.12.0",
+    "ruff>=0.1.9",
+    "mypy>=1.8.0",
+]
+
+[project.scripts]
+loki = "loki.cli:main"
+
+[tool.pytest.ini_options]
+minversion = "7.0"
+testpaths = ["tests"]
+python_files = ["test_*.py"]
+python_classes = ["Test*"]
+python_functions = ["test_*"]
+addopts = [
+    "-v",
+    "--strict-markers",
+    "--tb=short",
+    "--disable-warnings",
+]
+markers = [
+    "integration: integration tests requiring model downloads",
+    "slow: tests that take significant time to run",
+]
+
+[tool.coverage.run]
+source = ["src/loki"]
+omit = [
+    "*/tests/*",
+    "*/__pycache__/*",
+    "*/venv/*",
+    "*/.venv/*",
+]
+
+[tool.coverage.report]
+exclude_lines = [
+    "pragma: no cover",
+    "def __repr__",
+    "raise AssertionError",
+    "raise NotImplementedError",
+    "if __name__ == .__main__.:",
+    "if TYPE_CHECKING:",
+    "@abstractmethod",
+]
+
+[tool.mypy]
+python_version = "3.12"
+warn_return_any = true
+warn_unused_configs = true
+disallow_untyped_defs = false
+ignore_missing_imports = true
+
+[[tool.mypy.overrides]]
+module = "tests.*"
+ignore_errors = true
+
+[tool.black]
+line-length = 100
+target-version = ['py312']
+include = '\.pyi?$'
+extend-exclude = '''
+/(
+  # directories
+  \.eggs
+  | \.git
+  | \.hg
+  | \.mypy_cache
+  | \.tox
+  | \.venv
+  | build
+  | dist
+)/
+'''
+
+[tool.ruff]
+line-length = 100
+target-version = "py312"
+
+[tool.ruff.lint]
+select = [
+    "E",  # pycodestyle errors
+    "W",  # pycodestyle warnings
+    "F",  # pyflakes
+    "I",  # isort
+    "B",  # flake8-bugbear
+    "C4",  # flake8-comprehensions
+    "UP",  # pyupgrade
+]
+ignore = [
+    "E501",  # line too long (handled by black)
+    "B008",  # do not perform function calls in argument defaults
+    "C901",  # too complex
+    "W293",  # blank line contains whitespace (cosmetic, handled by black)
+    "UP007", # Use X | Y syntax for Union (prefer Union for complex type aliases)
+]
+
+[tool.ruff.lint.per-file-ignores]
+"__init__.py" = ["F401"]
+"tests/*" = ["F401", "F811"]
+"scripts/*" = ["E402"]  # Module imports after path manipulation
+
+[tool.isort]
+profile = "black"
+line_length = 100
